requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
Pillow==10.1.0
python-dotenv==0.21.1
requests==2.31.0
tk==0.1.0
urllib3==2.1.0
